Rename "walker" to "plainer"
